+++ OUDS/Core/Themes/Sosh/Sources/SoshTheme.swift
@@ -160,9 +160,11 @@ extension Tuning {
     /// The theme tuning for *Sosh *contains:
     /// - square corners for buttons
     /// - square corners for text inputs
+    /// - square corners for PIN code inputs
     /// - rounded corners for alert messages
     public static let Sosh = Tuning(hasRoundedButtons: false,
                                     hasRoundedTextInputs: false,
+                                    hasRoundedPinCodeInputs: false,
                                     hasRoundedAlertMessages: true)
 }