+++ OUDS/Core/Themes/Wireframe/Sources/WireframeTheme.swift
@@ -156,9 +156,11 @@ extension Tuning {
     /// The theme tuning for *Wireframe *contains:
     /// - square corners for buttons
     /// - square corners for text inputs
+    /// - square corners for PIN code inputs
     /// - rounded corners for alert messages
     public static let Wireframe = Tuning(hasRoundedButtons: false,
                                          hasRoundedTextInputs: false,
+                                         hasRoundedPinCodeInputs: false,
                                          hasRoundedAlertMessages: true)
 }